content/media/test/test_playback.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test playback of media files that should play OK</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var manager = new MediaTestManager;

function startTest(test, token) {
  var v = document.createElement('video');
  v.token = token;
  manager.started(token);

  v.src = test.name;
  v.name = test.name;
  var check = function(test, v) { return function() {
    is(test.name, v.name, "Name should match test.name #1");
    checkMetadata(test.name, v, test);
  }}(test, v);
  var noLoad = function(test, v) { return function() {
    ok(false, test.name + " should not fire 'load' event");
  }}(test, v);
  var checkEnded = function(test, v) { return function() {
    if (test.duration) {
      ok(Math.abs(v.currentTime - test.duration) < 0.1,
         test.name + " current time at end: " + v.currentTime + " should be: " + test.duration);
    }
    is(test.name, v.name, "Name should match test.name #2");
    is(v.readyState, v.HAVE_CURRENT_DATA, test.name + " checking readyState");
    ok(v.readyState != v.NETWORK_LOADED, test.name + " shouldn't report NETWORK_LOADED");
    ok(v.ended, test.name + " checking playback has ended");
    if (v.ended && v.seenSuspend && !v.finished) {
      v.finished = true;
      v.parentNode.removeChild(v);
      manager.finished(v.token);
    }
  }}(test, v);
  var checkSuspended = function(test, v) { return function() {
    is(test.name, v.name, "Name should match test.name #3");
    if (v.seenSuspend)
      return;

    v.seenSuspend = true;
    ok(true, test.name + " got suspend");
    if (v.ended && !v.finished) {
      v.finished = true;
      v.parentNode.removeChild(v);
      manager.finished(v.token);
    }
  }}(test, v);
  v.prevTime = 0;
  var timeUpdate = function(test, v) { return function() {
    is(test.name, v.name, "Name should match test.name #4");
    checkMetadata(test.name, v, test);
    ok(v.prevTime <= v.currentTime,
       test.name + " time should run forwards: p=" +
       v.prevTime + " c=" + v.currentTime);
    v.prevTime = v.currentTime;
  }}(test, v);
  v.addEventListener("load", noLoad, false);
  v.addEventListener("loadedmetadata", check, false);
  v.addEventListener("timeupdate", timeUpdate, false);

  // We should get "ended" and "suspend" events for every resource
  v.addEventListener("ended", checkEnded, false);
  v.addEventListener("suspend", checkSuspended, false);

  document.body.appendChild(v);
  v.play();
}

manager.runTests(gPlayTests, startTest);

</script>
</pre>
</body>
</html>
